google: add bucket component
doc: https://www.terraform.io/docs/providers/google/r/storage_bucket.html

To add this component, some updates have been made:

* add API to template

since the beginning, we work only with `compute` APIs. We need to
be able to easily switch from one API to the other when we declare a new
component.
In order to keep compatibility, default API goes to `compute`

* add `NoFilter` option to the template

most of the resource are based on filters. We need a way to
specify if a resource is not using filter

* add `storage` to the reader

* add `ResourceList` option

ResourceList overrides the default name of the resources list: `resourceList`
exemple:
for the components Instance, the list struct is `InstanceList`
but for the components Bucket, the list struct is `Buckets`