provider: Changed the core logic to not rely the old code and use the GRPC Clients
TF uses internally GRPC to communicate with other providers, this is not something we want to do as we want this to be stand alone lib
but we followed the same structure on how to do that so we could reuse more code from the Providers and TF.

The main issue is that they do not have the same types on compile time and the GRPC client cannot be reused from TF so we had to make a custom
copy which can support any Provider.

Whith this implementation we can use the SDK-V2 and have more reusable code into Client abstraction.